Adding Files for Difficult Project
The search results page, which is the optional 'difficult' HTML/CSS project for Web Development 101, will really only need one html file, one css file, and then some pictures for some of the other elements of the page.

The purpose of this commit is solely to add the html and css files (pictures will be added later).