Adding elements necessary for top-line navigation
Created structure for the header section, such as search bar elements (microphone and magnifying glass). This will allow for simply formatting to get the different elements in place. The overall methodology (for this entire webpage) being to get one area fully formatted out before moving on to the next part.