Create Sample Results w/ Minor Formatting
Here I just took the second result from the original results page and used it as a filler as I format everything. I copied it 6 times to make it reflect the number of results on the real page. I'll continue to use these samples until I'm done with formatting and then I'll replace their contents with the rest of the actual results. Some minor formatting was done on these just to test out the idea. It should work well.