Format search results
In this session, I finished formatting the normal results that I had created before and then I created and formatted the first search result, which is organized a bit differently than the other results on the page (since it is made of a table view instead of just aligned elemnts). It was good fun, and I got a quick refresher on tables.